+++ src/Enums/XLengthUnit.php
@@ -5,7 +5,7 @@ namespace gecleanme\Xkunverio\Enums;
 /**
  * Enum representing different length units and their conversion factors.
  */
-enum XLenUnit: string
+enum XLengthUnit: string
 {
     /**
      * Inch unit.
@@ -64,5 +64,15 @@ enum XLenUnit: string
             self::Meter => 1,
             self::Kilometer => 1000,
         };
+    }
+
+    /**
+     * Get Enum case values as an array.
+     *
+     * @return array Enum case values array.
+     */
+    public static function getCases(): array
+    {
+        return array_column(self::cases(), 'value');
     }
 }