+++ src/Enums/XVolumeUnit.php
@@ -2,7 +2,7 @@
 
 namespace gecleanme\Xkunverio\Enums;
 
-enum XVolUnit: string
+enum XVolumeUnit: string
 {
     /**
      * Cubic Inches unit.