+++ CHANGELOG.md
@@ -7,3 +7,12 @@ All notable changes to `xkunverio` will be documented in this file.
 
 # 1.0.1
 - Added DocBlocks
+
+# 2.1.0
+- Added Support for the following conversion types:
+  - Area
+  - Mass
+  - Speed
+  - Volume
+  - Temperature
+- Completed Tests